added place data in eventslicei and uisng that for place holder in autocomplete
