+++ .eslintrc
@@ -57,8 +57,9 @@
       }
     ],
     "react/jsx-key": "error",
+    // some problem with react project, sometimes it keeps showing error even if nothing extraneous is imported
     "import/no-extraneous-dependencies": [
-      "warn",
+      "error",
       {
         "devDependencies": [
           "**/*.test.js",
@@ -92,10 +93,8 @@
     // the stateless function rule is unusually restrict
     "react/prefer-stateless-function": "off",
     // unusually retrict
-    "react/static-property-placement": "off",
-    "import/no-named-as-default": "off"
-    // some problem with react project
-
+    "react/static-property-placement": "off"
+    // some problem with react project, sometimes file changes do not get reflected immedidately by typescript and it keeps showing errors
   }
 }
 // no undersocre dangle had to be turned off because _id of business needed to be accessed which was throwing errors
