+++ public/index.html
@@ -10,9 +10,11 @@
       content="Web site created using create-react-app"
     />
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
-    <link rel="stylesheet" href="%PUBLIC_URL%/fonts/fonts.css">
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" />
-
+    <link rel="stylesheet" href="%PUBLIC_URL%/fonts/fonts.css" />
+    <link
+      rel="stylesheet"
+      href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap"
+    />
 
     <!--
       manifest.json provides metadata used when your web app is installed on a
@@ -33,6 +35,7 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
+    <!--    this is a google map requirement, a div with id map should be present for maps to render correctly-->
     <div id="map"></div>
 
     <!--
