+++ src/App.scss
@@ -16,10 +16,4 @@ html {
 }
 .App {
 }
-.MuiInput-underline {
-  border: none !important;
-}
 
-.MuiInputBase-root .InputBase-adornedStart .MuiInputBase-adornedEnd {
-  height: 100% !important;
-}
