+++ src/features/event/EventMap.tsx
@@ -32,12 +32,12 @@ class EventMap extends React.Component<IProps, IState> {
         streetViewControl={false}
         fullscreenControl={false}
         initialCenter={{
-          lat: this.props.markerLocation.lat,
-          lng: this.props.markerLocation.lng,
+          lat: this.props.markerLocation.lat || config.SAN_FRANCISCO_LAT,
+          lng: this.props.markerLocation.lng || config.SAN_FRANCISCO_LONG,
         }}
         center={{
-          lat: this.props.markerLocation.lat,
-          lng: this.props.markerLocation.lng,
+          lat: this.props.markerLocation.lat || config.SAN_FRANCISCO_LAT,
+          lng: this.props.markerLocation.lng || config.SAN_FRANCISCO_LONG,
         }}
       >
         {/* @ts-ignore */}
@@ -46,8 +46,8 @@ class EventMap extends React.Component<IProps, IState> {
           name="Selected Location"
           // @ts-ignore
           position={{
-            lat: this.props.markerLocation.lat,
-            lng: this.props.markerLocation.lng,
+            lat: this.props.markerLocation.lat || config.SAN_FRANCISCO_LAT,
+            lng: this.props.markerLocation.lng || config.SAN_FRANCISCO_LONG,
           }}
         />
       </Map>
